requirements.txt
requests==2.24.0
beautifulsoup4==4.9.1
selenium==3.141.0
SQLAlchemy==1.3.17
psycopg2==2.8.5
python-dotenv~=0.13.0